formulaires de recensements personnalisés - fix #98
