+++ shared/db/migrate/20220615093906_add_recensement_forms_pdf_to_communes.rb
@@ -1,0 +1,5 @@
+class AddRecensementFormsPdfToCommunes < ActiveRecord::Migration[7.0]
+  def change
+    add_column :communes, :recensement_forms_pdf_updated_at, :datetime
+  end
+end
